Move all utils.general tests into separate folder
